requirements.txt
openai>=1.0.0
sentence-transformers>=2.2.2
pandas>=2.0.0
Pillow>=10.0.0
requests>=2.31.0
tqdm>=4.65.0
imgurpython>=1.1.7
python-dotenv>=1.0.0 